feat: rename host automation & browser extension
