+++ packages/browser-extension/README.md
