+++ packages/browser-extension/package.json
