+++ packages/browser-extension/sidepanel.html
