+++ packages/browser-extension/src/chatkit-options.test.ts
