+++ packages/browser-extension/src/chatkit-options.ts
