+++ packages/browser-extension/src/config.test.ts
