+++ packages/browser-extension/src/config.ts
