+++ packages/browser-extension/src/host.ts
