+++ packages/browser-extension/src/i18n.test.ts
