+++ packages/browser-extension/src/i18n.ts
