+++ packages/browser-extension/src/index.ts
