+++ packages/browser-extension/src/options.ts
