+++ packages/browser-extension/src/platform/chrome/api.test.ts
