+++ packages/browser-extension/src/platform/chrome/manifest.test.ts
