+++ packages/browser-extension/src/platform/chrome/manifest.ts
