+++ packages/browser-extension/src/popup.ts
