+++ packages/browser-extension/src/service-worker.ts
