+++ packages/browser-extension/src/sidepanel.ts
