+++ packages/browser-extension/tsconfig.json
