+++ packages/browser-extension/vite.config.ts
