+++ packages/browser-extension/vitest.config.ts
