+++ packages/host-automation/package.json
