+++ packages/host-automation/src/types.ts
