+++ packages/host-automation/tsconfig.json
