+++ packages/host-automation/vite.config.ts
