+++ packages/host-automation/vitest.config.ts
